extracted AllTasksView
