+++ TaskManagement/TaskManagement/PresentationLayer/Modules/Adding/AddingViewModel.swift
@@ -10,4 +10,12 @@ class AddingViewModel: ObservableObject {
     @Published var taskDate: Date = .now
     @Published var taskCategory: TaskCategory = .normal
     @Published var shouldSendNotification: Bool = false
+
+    func reset() {
+        self.taskTitle = ""
+        self.taskDescription = ""
+        self.taskDate = .now
+        self.shouldSendNotification = false
+        self.taskCategory = .normal
+    }
 }
